Use proxrox to define proxy strategy
The movie database currently relies on manual nginx configuration. This
means that developers need to copy the nginx config to system
directories and also keep them in sync with the files from the
repositories.

Additionally, the config file does not include instructions for GZip
and logs which might come in handy during development.

Proxrox is a viable alternative to manual nginx configuration as it
has support for automatic nginx installation (for OS X) and a config
file format that can be shared via the repository.